PLAT-1576 - support for variables in stage
implemented by using serverless/serverless#6566 which adds support for:
 * an `initialize` lifecycle hook
 * real support for adding variable getters via the `variableGetters` property
   on plugins

backwards compatibility is maintained by virtue of keeping the old way, which is
entirely configured in the `asyncInit` funciton. This does not cause forwards
compatibility problems because the above PR removes the call to `asyncInit` in
plugins as it was only ever meant for use with this plugin.