Emit an incomplete payload in NodeJS as timeout approaches
In NodeJS, a subprocess is `spawn`'d that sleeps for the duration of the
timeout then calls `kill $PPID` to send the main lambda process a
`SIGTERM`.

In Python, a `threading.Timer` is used to spawn a thread that waits
until the timeout, and then calls `SIGTERM` on the process.

In both, a `SIGTERM` handler is registered and upon recieving it, a
`timeout` transaction is written.

On succesfull exceution, in NodeJS, the subprocess is killed and, in