Fix AttributeError when error response None
The `error` can have a `response` attribute with the value `None`, but
the existing handling assumes any `error.response` is a `dict` (or at
least, has a `get` method). This causes an `AttributeError` when we try
to use it.

This commit fixes the error by setting `response = {}` when there is no
`error.response` _or_ it is `None`, rather than only in the former case.